requirements.txt
# Core
sqlalchemy
psycopg2-binary
python-dotenv

# API
fastapi
pydantic
uvicorn[standard]

# Normalization (HTML/PDF parsing)
beautifulsoup4
PyPDF2

# Embeddings
openai
sentence-transformers
numpy
torch  # sentence-transformers soft dependency
# rank-bm25
# fugashi
